Safari-611.1.10/LayoutTests/fast/text/font-face-set-ready-fire-expected.txt

This test makes sure that document.fonts.ready always triggers its promise when there are no loads.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.fonts.status is "loaded"
PASS document.fonts.status is "loading"
PASS successfullyParsed is true

TEST COMPLETE
